Use develop mode instead of building in tox
When using watchexec[1] to run tox on change events, I noticed that it
would continually loop. Placing some debugging lines into that tool, I
found that the file `solitatire_player.egg-info/PKG-INFO` file was
registering as changed all the time, despite the directory being in the
gitignore.

Using the develop mode skips creating this, and allows watchexec to work
as expected.

[1] https://github.com/mattgreen/watchexec